Added confirmations

+++ src/store/services/model-service.js
@@ -1,7 +1,17 @@
 import { axios } from '@/axios';
 import { store as storage } from '@/store';
-import Swal from 'sweetalert2';
-// import router from "@/router";
+import $Swal from 'sweetalert2';
+const $swal = $Swal.mixin({
+    customClass: {
+        confirmButton: 'btn btn-primary',
+        cancelButton: 'btn btn-outline-primary'
+    },
+    buttonsStyling: false,
+    cancelButtonText: "Cancelar"
+})
+
+import { createToaster } from "@meforma/vue-toaster";
+const toaster = createToaster({ /* options */ });
 
 const url = process.env.VUE_APP_API_BASE_URL;
 
@@ -52,7 +62,7 @@ function store(model, payload, multipart = false) {
 
     storage.commit('loading');
     return axios.post(`${url}/${apis[model]}`, payload, options).then((response) => {
-        Swal.fire('Registro completado', 'El registro se ha completado correctamente', 'success')
+        $swal.fire('Registro completado', 'El registro se ha completado correctamente', 'success')
         storage.commit('loading');
         return response
     }).catch(err => {
@@ -96,23 +106,44 @@ function show(model, id, params = "") {
     });
 }
 
-function destroy(model, id) {
+async function destroy(model, id) {
     storage.commit('loading');
-    return axios.delete(`${url}/${apis[model]}/${id}`).then((response) => {
-        Swal.fire('Registro eliminado', 'El registro se ha eliminado correctamente', 'success')
-        storage.commit('loading');
-        return response
-    }).catch(err => {
-        storage.commit('loading');
-        const status = err.response.status
-
-        let message = null;
-        if (status == 422) {
-            message = err.response.data.message
+    let deletes = false
+    await $swal.fire({
+        title: "¿Estas seguro de esto?",
+        text: "Una vez eliminado no podra recuperarse este registro",
+        icon: "warning",
+        showCancelButton: true,
+        showLoaderOnConfirm: true
+    }).then((res) => {
+        if (res.isConfirmed) {
+            deletes = true
+        } else {
+            deletes = false
         }
-        errors(status, message)
-        throw Error('Error');
-    });;
+    })
+
+    if (deletes) {
+        return axios.delete(`${url}/${apis[model]}/${id}`).then((response) => {
+            storage.commit('loading');
+            toaster.success('Registro eliminado')
+            return response
+        }).catch(err => {
+            storage.commit('loading');
+            const status = err.response.status
+
+            let message = null;
+            if (status == 422) {
+                message = err.response.data.message
+            }
+            errors(status, message)
+            toaster.error('No se pudo eliminar el registro')
+            throw Error('Error');
+        })
+    } else {
+        storage.commit('loading');
+        return null
+    }
 }
 
 function instByBusiness(id, page = 1) {
@@ -187,24 +218,42 @@ function users(params) {
     });;
 }
 
+function dashboard() {
+    storage.commit('loading');
+    return axios.get(url + '/dashboard').then((response) => {
+        storage.commit('loading');
+        return response
+    }).catch(err => {
+        storage.commit('loading');
+        const status = err.response.status
+
+        errors(status)
+        throw Error('Error');
+    });;
+}
+
 function errors(code, message = null) {
     switch (code) {
         case 500:
             // router.back()
-            Swal.fire('Error', message != null ? message : 'Parece que ocurrio un error en el servidor, porfavor intentalo mas tarde.', 'error')
+            $swal.fire('Error', message != null ? message : 'Parece que ocurrio un error en el servidor, porfavor intentalo mas tarde.', 'error')
             break;
 
         case 404:
-            Swal.fire('No encontrado', message != null ? message : 'Parece que ocurrio un error al intentar entontrar el recurso solicitado.', 'error')
+            $swal.fire('No encontrado', message != null ? message : 'Parece que ocurrio un error al intentar entontrar el recurso solicitado.', 'error')
             break;
 
         case 422:
-            Swal.fire('Algunos campos son incorrectos', message != null ? message : "Algunos de los datos enviados son incorrectos.", 'error')
+            $swal.fire('Algunos campos son incorrectos', message != null ? message : "Algunos de los datos enviados son incorrectos.", 'error')
             break;
+
+        case 401:
+            window.location.reload
+            break
 
         default:
             // router.back()
-            Swal.fire('Error inesperado', message != null ? message : 'Ocurrio un error inesperado', 'error')
+            $swal.fire('Error inesperado', message != null ? message : 'Ocurrio un error inesperado', 'error')
             break;
     }
 }
@@ -218,5 +267,6 @@ export default {
     update,
     getDocument,
     users,
-    errors
+    errors,
+    dashboard
 }
